Changed dtype to usize

+++ crates/polars-ops/src/chunked_array/cov.rs
@@ -13,7 +13,7 @@ fn multiply_sum(x: &[f64; COV_BUF_SIZE], y: &[f64; COV_BUF_SIZE], k: usize) -> f
 }
 
 /// Compute the covariance between two columns.
-pub fn cov<T>(a: &ChunkedArray<T>, b: &ChunkedArray<T>, ddof: u8, min_periods: u8) -> Option<f64>
+pub fn cov<T>(a: &ChunkedArray<T>, b: &ChunkedArray<T>, ddof: u8, min_periods: usize) -> Option<f64>
 where
     T: PolarsNumericType,
     T::Native: ToPrimitive,
@@ -44,7 +44,7 @@ where
 
 /// # Arguments
 /// `iter` - Iterator over `T` tuple where any `Option<T>` would skip the tuple.
-fn online_cov<I, J, T>(iters: I, ddof: u8, min_periods: u8) -> f64
+fn online_cov<I, J, T>(iters: I, ddof: u8, min_periods: usize) -> f64
 where
     I: Iterator<Item = J>,
     J: IntoIterator<Item = (T, T)> + Clone,
@@ -102,7 +102,7 @@ where
 }
 
 /// Compute the pearson correlation between two columns.
-pub fn pearson_corr<T>(a: &ChunkedArray<T>, b: &ChunkedArray<T>, ddof: u8, min_periods: u8) -> Option<f64>
+pub fn pearson_corr<T>(a: &ChunkedArray<T>, b: &ChunkedArray<T>, ddof: u8, min_periods: usize) -> Option<f64>
 where
     T: PolarsNumericType,
     T::Native: ToPrimitive,
@@ -130,7 +130,7 @@ where
 
 /// # Arguments
 /// `iter` - Iterator over `T` tuple where any `Option<T>` would skip the tuple.
-fn online_pearson_corr<I, J, T>(iters: I, ddof: u8, min_periods: u8) -> f64
+fn online_pearson_corr<I, J, T>(iters: I, ddof: u8, min_periods: usize) -> f64
 where
     I: Iterator<Item = J>,
     J: IntoIterator<Item = (T, T)> + Clone,
